% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{draw.bar}
\alias{draw.bar}
\title{Plot a colour bar.}
\usage{
draw.bar(colornames, texts = TRUE)
}
\arguments{
\item{colornames}{A vector of colour names.}

\item{texts}{Whether to show text labels of colour names.}
}
\value{
Draw a colour bar.
}
\description{
Plot a colour bar.
}
