% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{get.imglist}
\alias{get.imglist}
\title{Get image url quote list from Google.}
\usage{
get.imglist(qword)
}
\arguments{
\item{qword}{String. Image search key word.}
}
\value{
An image url quote list with 20 entries. 
Google limits number of images returned by 20.
}
\description{
Get image url quote list from Google.
}

