% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{hunt.colors}
\alias{hunt.colors}
\title{Extract colours from Google image search results.}
\usage{
hunt.colors(qword, num = 5, index = 1, seed = 42, plot = TRUE)
}
\arguments{
\item{qword}{String. Image search key word.}

\item{num}{Number of colours to extract.}

\item{index}{An integer or a vector indicating which image(s) to use.
Images can be viewed using \code{show.imgs}.}

\item{seed}{Seed for k-means.}

\item{plot}{Whether to plot the colours.}
}
\value{
A vector of colour names.
}
\description{
Extract colours from Google image search results.
}
\examples{
show.imgs("stardew valley")
hunt.colors("stardew valley", 10, 1:10)
}
\seealso{
\code{\link{show.imgs}}.
}

