% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{import}
\alias{import}
\title{Import one image from an url quote, output 2d or 3d array.}
\usage{
import(imgquote, out2d = FALSE)
}
\arguments{
\item{imgquote}{image url quote generated by \code{\link{get.imglist}}.}

\item{out2d}{If \code{TRUE}, the 3D array of the image will be
transformed into 2D.}
}
\value{
A 2D or 3D array representing the image.
}
\description{
Import one image from an url quote, output 2d or 3d array.
}
