% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{k.colors}
\alias{k.colors}
\title{Find main colours by k-means.}
\usage{
k.colors(img, num = 5, seed = 42)
}
\arguments{
\item{img}{A 2D or 3D array. 3D arrays will be transformed into 2D.}

\item{num}{Number of colours to extract.}

\item{seed}{Seed for k-means.}
}
\description{
Find main colours by k-means.
}
