% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{make.gradient}
\alias{make.gradient}
\title{Make a colour gradient.}
\usage{
make.gradient(..., len = 5, fixed = NULL, order = FALSE, plot = TRUE,
  texts = TRUE)
}
\arguments{
\item{...}{Color names.}

\item{len}{Length of the gradient. If more than two colors, the length of
the step (A 4-colour gradient with \code{len = 5} gives 15 colours).}

\item{fixed}{A concatenated string indicating the fixed component(s) of 
the gradient. One or two components (letters) of either \code{"rgb"} or 
\code{"hsv"} can be specified. See details.}

\item{order}{If \code{TRUE}, colours are ordered to make a continuous
gradient from dark to light.}

\item{plot}{Whether to plot the gradient bar. Default \code{TRUE}.}

\item{texts}{Whether to add text labels to the bar. Default \code{TRUE}}
}
\value{
A vector of colour names.
}
\description{
Make a colour gradient.
}
\details{
Examples of valid \code{fixed} are \code{"hv", "h", "rg"} etc.
Two-colour gradient cannot be used in combination with \code{fixed}.
If only one colour is specified, the colour's position in the gradient
is decided by its brightness (if \code{fixed} is \code{NULL}) or by its
value of the first fixed component (if \code{fixed} is specified).
}
\examples{
make.gradient()
make.gradient("turquoise", fixed = "hv")
make.gradient(cyans[1], blues[9], len = 8)
make.gradient(blues[1], reds[1], yellows[1], cyans[1])
}
