% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{pick.n}
\alias{pick.n}
\title{Pick n most distinctive colours out of many colours.}
\usage{
pick.n(colornames, n, plot = FALSE)
}
\arguments{
\item{colornames}{A vetor of colour names.}

\item{n}{How many colours to pick.}

\item{plot}{Whether to plot the colour bar. Default \code{FALSE}.}
}
\value{
Picked colour names (ordered from dark to light).
}
\description{
Pick n most distinctive colours out of many colours.
}
\details{
Makes the process of choosing colours easier especially when 
choosing legend colours in a list of gradient theme colours.
}
\examples{
pink.n(make.gradient(occblues, len = 2), len = 8, plot = TRUE)
}

