% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{rects}
\alias{rects}
\title{Draw rectangles/ grids.}
\usage{
rects(xl, yt, w, h, ncol = 3, nrow = 1, border = "transparent",
  add = FALSE, col = "white", texts = NULL, texts.col = "auto", ...)
}
\arguments{
\item{xl}{First/ left-most xleft.}

\item{yt}{First/ top-most ytop.}

\item{w}{Width of rectangles.}

\item{h}{Height of rectangles.}

\item{ncol}{Number of columns.}

\item{nrow}{Number of rows. Default 1.}

\item{border}{Default to transparent.}

\item{add}{Default to \code{FALSE}. The function detects whether a plot
device is active and if not, calls plot.new before plotting. If a plot 
has been called and rects should be added to plot, set to \code{TRUE}.}

\item{col}{Fill colours of rects.}

\item{texts}{Text labels in the rectangles/ grid. Order is by column.}

\item{texts.col}{Can be colours or one of the two keywords \code{"auto"}
and \code{"theme"}. \code{"auto"} would make black/ white texts based on
the underlying rectangle colour. \code{"theme"} would choose the colours
from the colours supplied to rectangles. Darkest one would be the black
equivalent and lightest one the white equivalent. Works well on one-
colour gradients.}
}
\value{
Draw rectangles.
}
\description{
Draw rectangles/ grids.
}
\examples{
rects(0, 5, 1, 1, 5, 5, col = occblues)
}

