% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{setup.null}
\alias{setup.null}
\title{Set up a null plot.}
\usage{
setup.null(xl0 = 0, xl1 = 5, yl0 = 0, yl1 = 5)
}
\arguments{
\item{xl0, xl1}{xlim. Default c(0, 5).}

\item{yl0, yl1}{ylim. Default c(0, 5).}
}
\description{
Set up a null plot.
}
