% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcolors.R
\name{show.imgs}
\alias{show.imgs}
\title{View 20 Google image search results.}
\usage{
show.imgs(qword)
}
\arguments{
\item{qword}{String. Image search key word.}
}
\description{
View 20 Google image search results.
}
\details{
Google limits number of images returned by 20.
}
\examples{
show.imgs("stardew valley")
hunt.colors("stardew valley", 10, 1:10)
}
\seealso{
\code{\link{hunt.colors}}
}
